Changing package to TopoChains.jl
While registering, noted that Stacks.jl is too general of a name (not to mention a Stack is an elementary data structure in CS literature, leading to potential name conflicts later down the line)

The TopoChain naming emphasizes this to be a drop-in replacement for a Chain, that just happens to store information abut the topology of a function composition.

Plus, new tests for the TopoChain structure to test compatibility with Flux layers.